refactor(watchStore): export from utils
