+++ src/vtk/ToolWidgetUtils/pointState.js
@@ -2,26 +2,10 @@ import macro from '@kitware/vtk.js/macros';
 import vtkWidgetState from '@kitware/vtk.js/Widgets/Core/WidgetState';
 import visibleMixin from '@kitware/vtk.js/Widgets/Core/StateBuilder/visibleMixin';
 import scale1Mixin from '@kitware/vtk.js/Widgets/Core/StateBuilder/scale1Mixin';
+import { watchStore } from '@/src/vtk/ToolWidgetUtils/utils';
 import { ANNOTATION_TOOL_HANDLE_RADIUS } from '@/src/constants';
 
 const PIXEL_SIZE = ANNOTATION_TOOL_HANDLE_RADIUS * 2;
-
-function watchStore(publicAPI, store, getter, cmp) {
-  let cached = getter();
-  const unsubscribe = store.$subscribe(() => {
-    const val = getter();
-    if (cmp ? cmp(cached, val) : cached !== val) {
-      cached = val;
-      publicAPI.modified();
-    }
-  });
-
-  const originalDelete = publicAPI.delete;
-  publicAPI.delete = () => {
-    unsubscribe();
-    originalDelete();
-  };
-}
 
 function _createPointState(
   publicAPI,
